requirements.txt
Flask
Flask-Cors
Flask-DotEnv
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
gunicorn
keras
nltk
numpy
python-dotenv
SQLAlchemy
tensorflow
tflearn